+++ packages/messaging/ios/RNFBMessaging/RNFBMessagingAppDelegateInterceptor.m
@@ -1,0 +1,81 @@
+/**
+ * Copyright (c) 2016-present Invertase Limited & Contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this library except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+#import <React/RCTConvert.h>
+#import <RNFBApp/RNFBSharedUtils.h>
+#import <RNFBApp/RNFBRCTEventEmitter.h>
+#import <GoogleUtilities/GULAppDelegateSwizzler.h>
+
+#import "RNFBMessagingAppDelegateInterceptor.h"
+#import "RNFBMessagingSerializer.h"
+
+@implementation RNFBMessagingAppDelegateInterceptor
+
++ (instancetype)sharedInstance {
+  static dispatch_once_t once;
+  static RNFBMessagingAppDelegateInterceptor *sharedInstance;
+  dispatch_once(&once, ^{
+    sharedInstance = [[RNFBMessagingAppDelegateInterceptor alloc] init];
+    [GULAppDelegateSwizzler proxyOriginalDelegateIncludingAPNSMethods];
+    [GULAppDelegateSwizzler registerAppDelegateInterceptor:sharedInstance];
+  });
+  return sharedInstance;
+}
+
+// used to temporarily store a promise instance to resolve calls to `registerForRemoteNotifications`
+- (void)setPromiseResolve:(RCTPromiseResolveBlock)resolve andPromiseReject:(RCTPromiseRejectBlock)reject {
+  _registerPromiseResolver = resolve;
+  _registerPromiseRejecter = reject;
+}
+
+// called when `registerForRemoteNotifications` completes successfully
+- (void)application:(UIApplication *)application didRegisterForRemoteNotificationsWithDeviceToken:(NSData *)deviceToken {
+  if (_registerPromiseResolver != nil) {
+    _registerPromiseResolver(@([RCTConvert BOOL:@([UIApplication sharedApplication].isRegisteredForRemoteNotifications)]));
+    _registerPromiseResolver = nil;
+    _registerPromiseRejecter = nil;
+  }
+}
+
+// called when `registerForRemoteNotifications` fails to complete
+- (void)application:(UIApplication *)application didFailToRegisterForRemoteNotificationsWithError:(NSError *)error {
+  if (_registerPromiseRejecter != nil) {
+    [RNFBSharedUtils rejectPromiseWithNSError:_registerPromiseRejecter error:error];
+    _registerPromiseResolver = nil;
+    _registerPromiseRejecter = nil;
+  }
+}
+
+// APNS - only called on iOS versions less than v10 (which isn't officially supported, added here for ease later on)
+- (void)application:(UIApplication *)application didReceiveRemoteNotification:(NSDictionary *)userInfo {
+  // send message event for remote notifications that have also have data
+  if (userInfo[@"aps"] && ((NSDictionary *) userInfo[@"aps"]).count >= 1 && userInfo[@"aps"][@"content-available"]) {
+    [[RNFBRCTEventEmitter shared] sendEventWithName:@"messaging_message_received" body:[RNFBMessagingSerializer remoteMessageAppDataToDict:userInfo withMessageId:nil]];
+  }
+}
+
+// APNS - only called on iOS versions greater than or equal to v10 (this one is officially supported)
+- (void)application:(UIApplication *)application didReceiveRemoteNotification:(NSDictionary *)userInfo fetchCompletionHandler:(void (^)(UIBackgroundFetchResult result))completionHandler {
+  // send message event for remote notifications that have also have data
+  if (userInfo[@"aps"] && ((NSDictionary *) userInfo[@"aps"]).count >= 1 && userInfo[@"aps"][@"content-available"]) {
+    [[RNFBRCTEventEmitter shared] sendEventWithName:@"messaging_message_received" body:[RNFBMessagingSerializer remoteMessageAppDataToDict:userInfo withMessageId:nil]];
+    // complete immediately
+    completionHandler(UIBackgroundFetchResultNoData);
+  }
+}
+
+@end